arcvault-triage/requirements.txt
langgraph>=0.2.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
gspread>=6.0.0
google-auth>=2.0.0
gradio>=4.0.0
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn>=0.30.0
pytest>=8.0.0
httpx>=0.27.0
